VectorCAN: Fix CAN-FD frame reading and sending
When setting the CAN device to flexible data rate mode the correct
sampling point has to be set.

When invalid CAN-FD configuration is set, respond with correct
error instead of always complain about open port.

When the bit rate switch is set for the sending frame, also set the
corresponding Vector flag.
Also set it to QCanBusFrame when receiving from Vector backend.

When an payload higher than 8 is used, the DLC has to be set to the
correct indexes.

Fixes: QTBUG-99259

+++ src/plugins/canbus/vectorcan/vectorcan_symbols_p.h
@@ -321,16 +321,16 @@ typedef struct {
 typedef struct {
      quint32 id;
      quint32 flags;
-     quint16 dlc;
-     quint16 reserved[7];
-     quint16 data[XL_CAN_MAX_DATA_LEN];
+     quint8 dlc;
+     quint8 reserved[7];
+     quint8 data[XL_CAN_MAX_DATA_LEN];
 } XL_CAN_TX_MSG;
 
 typedef struct {
-    quint32 tag;
-    quint32 transId;
-    quint32 channelIndex;
-    quint32 reserved[3];
+    quint16 tag;
+    quint16 transId;
+    quint8 channelIndex;
+    quint8 reserved[3];
     union {
         XL_CAN_TX_MSG canMsg;
     } tagData;
